requirements.txt
# MCP Architecture Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
structlog==23.2.0